Use dtype=float in tests that need float precision

+++ tests/test_detectors.py
@@ -294,7 +294,7 @@ def test_quantize():
 def test_IPC_basic():
     # Make an image with non-trivially interesting scale.
     g = galsim.Gaussian(sigma=3.7)
-    im = g.drawImage(scale=0.25)
+    im = g.drawImage(scale=0.25, dtype=float)
     im_save = im.copy()
 
     # Check for no IPC
